requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.13
python-dotenv==1.0.1
pydantic==2.8.2

streamlit==1.37.1
pandas==2.2.2

pytest==8.3.2
httpx==0.27.0
